Extract GrammarIssuesCollector from ANTLRv4ExternalAnnotator
