+++ src/main/java/org/antlr/intellij/plugin/ANTLRv4ExternalAnnotator.java
@@ -6,39 +6,24 @@ import com.intellij.lang.annotation.ExternalAnnotator;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.util.TextRange;
-import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiFile;
-import org.antlr.intellij.plugin.parsing.RunANTLROnGrammarFile;
 import org.antlr.intellij.plugin.psi.MyPsiUtils;
-import org.antlr.intellij.plugin.validation.GrammarInfoMessage;
 import org.antlr.intellij.plugin.validation.GrammarIssue;
-import org.antlr.runtime.ANTLRReaderStream;
+import org.antlr.intellij.plugin.validation.GrammarIssuesCollector;
 import org.antlr.runtime.CommonToken;
 import org.antlr.runtime.Token;
-import org.antlr.v4.Tool;
-import org.antlr.v4.parse.ANTLRParser;
-import org.antlr.v4.runtime.misc.IntervalSet;
-import org.antlr.v4.tool.*;
-import org.antlr.v4.tool.ast.GrammarAST;
-import org.antlr.v4.tool.ast.GrammarRootAST;
-import org.antlr.v4.tool.ast.RuleRefAST;
+import org.antlr.v4.tool.ErrorSeverity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.stringtemplate.v4.ST;
 
-import java.io.File;
-import java.io.StringReader;
-import java.util.*;
-
-import static org.antlr.v4.codegen.CodeGenerator.targetExists;
+import java.util.List;
 
 public class ANTLRv4ExternalAnnotator extends ExternalAnnotator<PsiFile, List<GrammarIssue>> {
     // NOTE: can't use instance var as only 1 instance
 
     public static final Logger LOG = Logger.getInstance("ANTLRv4ExternalAnnotator");
-	private static final String LANGUAGE_ARG_PREFIX = "-Dlanguage=";
 
-	/** Called first; return file */
+    /** Called first; return file */
 	@Override
 	@Nullable
 	public PsiFile collectInformation(@NotNull PsiFile file) {
@@ -49,98 +34,10 @@ public class ANTLRv4ExternalAnnotator extends ExternalAnnotator<PsiFile, List<Gr
 	@Nullable
 	@Override
 	public List<GrammarIssue> doAnnotate(final PsiFile file) {
-		String grammarFileName = file.getVirtualFile().getPath();
-		LOG.info("doAnnotate "+grammarFileName);
-		String fileContents = file.getText();
-		List<String> args = RunANTLROnGrammarFile.getANTLRArgsAsList(file.getProject(), file.getVirtualFile());
-		AnnotatorToolListener listener = new AnnotatorToolListener();
-
-		String languageArg = findLanguageArg(args);
-
-		if ( languageArg!=null ) {
-			String language = languageArg.substring(LANGUAGE_ARG_PREFIX.length());
-
-			if ( !targetExists(language) ) {
-				GrammarIssue issue = new GrammarIssue(null);
-				issue.setAnnotation("Unknown target language '" + language + "', analysis will be done using the default target language 'Java'");
-				listener.issues.add(issue);
-
-				args.remove(languageArg);
-			}
-		}
-
-		final Tool antlr = new Tool(args.toArray(new String[args.size()]));
-		if ( !args.contains("-lib") ) {
-			// getContainingDirectory() must be identified as a read operation on file system
-			ApplicationManager.getApplication().runReadAction(new Runnable() {
-				@Override
-				public void run() {
-					antlr.libDirectory = file.getContainingDirectory().toString();
-				}
-			});
-		}
-
-		antlr.removeListeners();
-        antlr.addListener(listener);
-		try {
-			StringReader sr = new StringReader(fileContents);
-			ANTLRReaderStream in = new ANTLRReaderStream(sr);
-			in.name = file.getName();
-			GrammarRootAST ast = antlr.parse(file.getName(), in);
-			if ( ast==null || ast.hasErrors ) {
-				for (GrammarIssue issue : listener.issues) {
-					processIssue(file, issue);
-				}
-				return listener.issues;
-			}
-			Grammar g = antlr.createGrammar(ast);
-			g.fileName = grammarFileName;
-
-			String vocabName = g.getOptionString("tokenVocab");
-			if ( vocabName!=null ) { // import vocab to avoid spurious warnings
-				LOG.info("token vocab file "+vocabName);
-				g.importTokensFromTokensFile();
-			}
-
-			VirtualFile vfile = file.getVirtualFile();
-			if ( vfile==null ) {
-				LOG.error("doAnnotate no virtual file for "+file);
-				return listener.issues;
-			}
-			g.fileName = vfile.getPath();
-			antlr.process(g, false);
-
-			Map<String, GrammarAST> unusedRules = getUnusedParserRules(g);
-			if ( unusedRules!=null ) {
-				for (String r : unusedRules.keySet()) {
-					Token ruleDefToken = unusedRules.get(r).getToken();
-					GrammarIssue issue = new GrammarIssue(new GrammarInfoMessage(g.fileName, ruleDefToken, r));
-					listener.issues.add(issue);
-				}
-			}
-
-			for (GrammarIssue issue : listener.issues) {
-				processIssue(file, issue);
-			}
-		}
-		catch (Exception e) {
-			LOG.error("antlr can't process "+file.getName(), e);
-		}
-		return listener.issues;
+		return GrammarIssuesCollector.collectGrammarIssues(file);
 	}
 
-	@Nullable
-	private String findLanguageArg(List<String> args) {
-		for ( String arg : args ) {
-			if ( arg.startsWith(LANGUAGE_ARG_PREFIX) ) {
-				return arg;
-			}
-		}
-
-		return null;
-	}
-
-	/** Called 3rd */
+    /** Called 3rd */
 	@Override
 	public void apply(@NotNull PsiFile file,
 					  List<GrammarIssue> issues,
@@ -201,60 +98,7 @@ public class ANTLRv4ExternalAnnotator extends ExternalAnnotator<PsiFile, List<Gr
 		super.apply(file, issues, holder);
 	}
 
-	public void processIssue(final PsiFile file, GrammarIssue issue) {
-		File grammarFile = new File(file.getVirtualFile().getPath());
-		if ( issue.getMsg() == null || issue.getMsg().fileName==null ) { // weird, issue doesn't have a file associated with it
-			return;
-		}
-		File issueFile = new File(issue.getMsg().fileName);
-		if ( !grammarFile.getName().equals(issueFile.getName()) ) {
-			return; // ignore errors from external files
-		}
-		ST msgST = null;
-		if ( issue.getMsg() instanceof GrammarInfoMessage ) { // not in ANTLR so must hack it in
-			Token t = ((GrammarSemanticsMessage) issue.getMsg()).offendingToken;
-			issue.getOffendingTokens().add(t);
-			msgST = new ST("unused parser rule <arg>");
-			msgST.add("arg", t.getText());
-			msgST.impl.name = "info";
-		}
-		else if ( issue.getMsg() instanceof GrammarSemanticsMessage ) {
-			Token t = ((GrammarSemanticsMessage) issue.getMsg()).offendingToken;
-			issue.getOffendingTokens().add(t);
-		}
-		else if ( issue.getMsg() instanceof LeftRecursionCyclesMessage ) {
-			List<String> rulesToHighlight = new ArrayList<String>();
-			LeftRecursionCyclesMessage lmsg = (LeftRecursionCyclesMessage) issue.getMsg();
-			Collection<? extends Collection<Rule>> cycles =
-				(Collection<? extends Collection<Rule>>)lmsg.getArgs()[0];
-			for (Collection<Rule> cycle : cycles) {
-				for (Rule r : cycle) {
-					rulesToHighlight.add(r.name);
-					GrammarAST nameNode = (GrammarAST)r.ast.getChild(0);
-					issue.getOffendingTokens().add(nameNode.getToken());
-				}
-			}
-		}
-		else if ( issue.getMsg() instanceof GrammarSyntaxMessage ) {
-			Token t = issue.getMsg().offendingToken;
-			issue.getOffendingTokens().add(t);
-		}
-		else if ( issue.getMsg() instanceof ToolMessage ) {
-			issue.getOffendingTokens().add(issue.getMsg().offendingToken);
-		}
-
-		Tool antlr = new Tool();
-		if ( msgST==null ) {
-			msgST = antlr.errMgr.getMessageTemplate(issue.getMsg());
-		}
-		String outputMsg = msgST.render();
-		if ( antlr.errMgr.formatWantsSingleLineMessage() ) {
-			outputMsg = outputMsg.replace('\n', ' ');
-		}
-		issue.setAnnotation(outputMsg);
-	}
-
-	public static String getFindVocabFileNameFromGrammarFile(PsiFile file) {
+    public static String getFindVocabFileNameFromGrammarFile(PsiFile file) {
 		final FindVocabFileRunnable findVocabAction = new FindVocabFileRunnable(file);
 		ApplicationManager.getApplication().runReadAction(findVocabAction);
 		return findVocabAction.vocabName;
@@ -274,25 +118,4 @@ public class ANTLRv4ExternalAnnotator extends ExternalAnnotator<PsiFile, List<Gr
 		}
 	}
 
-	public static Map<String,GrammarAST> getUnusedParserRules(Grammar g) {
-		if ( g.ast==null || g.isLexer() ) return null;
-		List<GrammarAST> ruleNodes = g.ast.getNodesWithTypePreorderDFS(IntervalSet.of(ANTLRParser.RULE_REF));
-		// in case of errors, we walk AST ourselves
-		// ANTLR's Grammar object might have bailed on rule defs etc...
-		Set<String> ruleRefs = new HashSet<String>();
-		Map<String,GrammarAST> ruleDefs = new HashMap<String,GrammarAST>();
-		for (GrammarAST x : ruleNodes) {
-			if ( x.getParent().getType()==ANTLRParser.RULE ) {
-//				System.out.println("def "+x);
-				ruleDefs.put(x.getText(), x);
-			}
-			else if ( x instanceof RuleRefAST ) {
-				RuleRefAST r = (RuleRefAST) x;
-//				System.out.println("ref "+r);
-				ruleRefs.add(r.getText());
-			}
-		}
-		ruleDefs.keySet().removeAll(ruleRefs);
-		return ruleDefs;
-	}
 }
